requirements.txt
streamlit
advertools
pandas
scikit-learn
google-generativeai